Remove redundant compose .env file
Since this file is used for starting
a compose workload inside a CI
pipeline, all key/value pairs found
in this file can be safely moved to
the docker-compose.yml file, thus
the compose maintainer needs to look
into one file only.
Another reason for removing this file
is that macOS-based agents use an
older version of Docker Compose
which seems it cannot handle the
`--env-file` CLI parameter.